Added createDeckTable function
The function receives a deck list (an object with the format: "card
name": cardCopies) and returns a table that is ready to be displayed.
The table will list the deck name/class, cards in the deck, how many of
the cards the user owns, how much dust is required to craft the
remaining cards of the deck, and the completion % in dust value. It will
list both normal and golden data, completion % is for combined data.