Added "Hide free cards" option
The option is located in the infobox and will hide free cards in the
first column. Common cards will be evenly split between the first two
columns instead. Many players already have every free card, so this
option was added to make more use of the first column for them, as well
as to reduce tracker length.